all: Fix 'mix signed and unsigned' CLang warnings

+++ src/public/materialsystem/imesh.h
@@ -964,7 +964,7 @@ inline void CVertexBuilder::Reset()
 
 	m_pCurrPosition = m_pPosition;
 	m_pCurrNormal = m_pNormal;
-	for ( int i = 0; i < NELEMS( m_pCurrTexCoord ); i++ )
+	for ( size_t i = 0; i < NELEMS( m_pCurrTexCoord ); i++ )
 	{
 		m_pCurrTexCoord[i] = m_pTexCoord[i];
 	}
